+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+# v3.3
+- New `fill_space!` function for discrete spaces.
+- The Daisyworld example now uses multi-agent approach (surface is agent).
+- New `allids` function.
+
+# v3.2
+- New `Models` submodule, that conveniently allows loading a model from the examples.
 # v3.1
-## Additions
 - Extend `interacting_pairs` to allow interactions of disparate types when using mixed models.
 
 # v3.0